Assignment 1
In this first programming assignment, which is the first part of the project  I wrote code to analyze the reading level of a piece of text. The Flesch Readability Score is a measure of the reading complexity of a piece of text. Developed by author Rudolf Flesch, it is a measure that approximates how easy a piece of text is to read based on the number of sentences, words and syllables in that text. Higher scores indicate text that is simple to read, while lower scores indicate more complex text.
Thus for this programming assignment I wrote the back-end code for calculating Flesch readability scores, and then integrated this back-end capability with the front-end application.