Assignment 3 : Part 2
So in this part I had to generate a bunch of words using the existing words that were typed in the text editor. This is called Markov text generation and it's based on the principle of training your text in which you use linked lists to keep track of the words that follow a particular word. After training you can genrate new sentences, using the lists and you have similar but new sentences generated!