Assignment 4
In this weeks Assignment I had to implement a Binary Search Tree (BST) and LinkedList(LL). I used Java's built in data structures, so I didn't have to write logic to implement them, but simply use them. After that I compared their run times to search a word in a BST and LL and boy was the BST way faster! I also had to implement a Trie which is similar to a BST to predict new words based on the half compelted word typed in, and Tries are very useful in doing so, they're actually faster BST's if the word length is smaller than 18 characters!