Assignment 5 : Part 1
In the first Part of the assignment, I picked up from last weeks assignment where I detected misspelled words and now I added functionality to my text editor by suggesting words when a user right clicks on the misspelled word. This was done by creating a mutation string which was just a bit different than the misspelled one, and from the bunch of mutaation strings, only the ones which were proper words were selcted and shown to the user.