Final Assignment : 5 Part 2
So with this part 2 of assignment 5, I've completed the course and I've finally built the smart text editor. In this assignment I had to calulate the Edit Distance from one word to the other and print the path that it took to get to the destination word from the source word. To do this I had to combine the logic of the first part of assignment 5 that is to get similar words and build a trie that kept track of it's parent and had it's own list of children. Once the word was found, I simply had to backtrack using the parents and print those words. It was a great assignment and I really enjoyed coding this one. So FINALLY I've finished the course and I've developed my own Smart Text Editor which is able to :
1. Load Text
2. Calculate Flech Score
3. Calculate Edit Distance
4. Generate Markov Text
5. Flag missplelled words
6. Show spelling suggestions by right clicking
7. show autocomplete for just a prefix of a word

It has truly been a great course and kudos to the UC San Diego MOOC Team for curating such a good course!